refactor(product list): Move files into product list template folder
