+++ frontend/components/common/index.ts
@@ -3,3 +3,4 @@ export * from './AppProviders';
 export * from './ProductCard';
 export * from './PageContentWrapper';
 export * from './SecondaryNavbar';
+export * from './useSearchCache';
