+++ frontend/components/common/useSearchCache.tsx
