+++ frontend/templates/product-list/ProductListDeviceNavigation.tsx
