+++ frontend/templates/product-list/ProductListView.tsx
