+++ frontend/templates/product-list/sections/BannerSection/index.tsx
