+++ frontend/templates/product-list/sections/FeaturedProductListSection.tsx
