+++ frontend/templates/product-list/sections/FilterableProductsSection/FacetsAccordion.tsx
