+++ frontend/templates/product-list/sections/FilterableProductsSection/ProductGrid.tsx
