+++ frontend/templates/product-list/sections/FilterableProductsSection/ProductList.tsx
