+++ frontend/templates/product-list/sections/FilterableProductsSection/RefinementSingleSelect.tsx
