+++ frontend/templates/product-list/sections/FilterableProductsSection/SearchInput.tsx
