+++ frontend/templates/product-list/sections/FilterableProductsSection/index.tsx
