+++ frontend/templates/product-list/sections/FilterableProductsSection/useCountRefinements.tsx
