+++ frontend/templates/product-list/sections/FilterableProductsSection/useDevicePartsItemType.tsx
