+++ frontend/templates/product-list/sections/FilterableProductsSection/useFacets.tsx
