+++ frontend/templates/product-list/sections/FilterableProductsSection/useFilteredRefinementList.tsx
