+++ frontend/templates/product-list/sections/FilterableProductsSection/useProductSearchHitPricing.ts
