+++ frontend/templates/product-list/sections/FilterableProductsSection/useSortBy.ts
