+++ frontend/templates/product-list/sections/HeroSection.tsx
