+++ frontend/templates/product-list/sections/ProductListChildrenSection.tsx
