+++ frontend/templates/product-list/sections/ProductListSetSection.tsx
