+++ frontend/templates/product-list/sections/RelatedPostsSection/PostCard.tsx
