+++ frontend/templates/product-list/sections/index.ts
